Safari-611.1.10/LayoutTests/workers/message-port-gc-expected.txt

Verify that collecting a transferred port should not interfere with posting messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
pong
pong
pong
pong
pong
pong
pong
pong
pong
pong
pong
pong
pong
pong
pong
pong
